cloud/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
